✨ Refactored Graph Structure & Node Separation.
